hls: Break out get_full_url tests to new test suite (http)
The get_full_url was moved some time ago from the hls module to the
svtplay_dl.utils.http module.
