+++ lib/svtplay_dl/tests/hls.py
@@ -10,7 +10,6 @@
 
 from __future__ import absolute_import
 import unittest
-import svtplay_dl.fetcher.hls as hls
 from svtplay_dl.fetcher.hls import M3U8
 
 # Example HLS playlist, source:
@@ -37,49 +36,6 @@ something0/else.m3u8
 
 
 class HlsTest(unittest.TestCase):
-    def test_get_full_url_1(self):
-        for test in [
-            # full http:// url as media segment in playlist
-            {
-                'srcurl': 'INVALID',
-                'segment': 'http://example.com/',
-                'expected': 'http://example.com/'
-            },
-            # full https:// url as media segment in playlist
-            {
-                'srcurl': 'INVALID',
-                'segment': 'https://example.com/',
-                'expected': 'https://example.com/'
-            },
-            # filename as media segment in playlist (http)
-            {
-                'srcurl': 'http://example.com/',
-                'segment': 'foo.ts',
-                'expected': 'http://example.com/foo.ts'
-            },
-            # filename as media segment in playlist (https)
-            {
-                'srcurl': 'https://example.com/',
-                'segment': 'foo.ts',
-                'expected': 'https://example.com/foo.ts'
-            },
-            # replacing srcurl file
-            {
-                'srcurl': 'http://example.com/bar',
-                'segment': 'foo.ts',
-                'expected': 'http://example.com/foo.ts'
-            },
-            # with query parameters
-            {
-                'srcurl': 'http://example.com/bar?baz=qux',
-                'segment': 'foo.ts',
-                'expected': 'http://example.com/foo.ts'
-            },
-        ]:
-            self.assertEqual(
-                hls._get_full_url(test['segment'], test['srcurl']),
-                test['expected'])
-
     def test_parse_m3u8(self):
         self.maxDiff = None
         for test in [
